remove gn.bbclass to gn-for-flutter.bbclass
gn.bbclass has the specific parts for flutter, so it is not generic,
then the name of "gn.bbclass" is misleading.
Therefore, I renamed this bbclass to "gn-for-flutter.bbclass".

For lib/gn.py, this fetcher can potentially fetch souce codes except flutter-engine
e.g. chromium, so I will not rename it.

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>