gn fetch: premirror compatibility
An attempt was successfully made at creating a fully reproducible tar
file for creating a premirror.  Unfortunately, testing showed that the
build eventually called git rev-parse, which then failed due to
stripping the .git directories out of the tar file.

So a different tack was taken: to accept that similar to git mirrors, gn
mirrors will always change and not be reproducible.  Code was primarily
borrowed from the git fetcher, but also the repo fetcher as it has some
similar characteristics.  The supports_checksum() function was added and
returns False now.