all: address do_compile proxy issues
The compile phase of certain packages depends on a loopback websocket
connection. If a proxy is exported for this without a corresponding
no_proxy config containing localhost, ::1, and 127.0.0.1, it will result
in a connection issue and error out.

If we're exporting most of the proxy variables manually now then lets
just use the export_proxy command from bitbake directly were proxies are
needed.

Signed-off-by: Randolph Sapp <rs@ti.com>