pdfium: fix building in arm64 hosts
'buildtools/reclient' in DEPS sets its package
name as 'package': Var('reclient_package') + '${{platform}}'.
This means that when building on arm64 hosts the package
name will solve to `infra/rbe/client/linux-arm64`,
which does not exist:

- https://chrome-infra-packages.appspot.com/p/infra_internal/rbe/client

The same issue was found in chromium and the solution was
to enable this package conditionally for non linux-arm64 hosts:

- https://issues.chromium.org/issues/40277110
- https://chromium.googlesource.com/chromium/src/+/d2c1621d8e02c804b67cf1722595754114cb9aa%5E%21/#F1

This commit adds a workaround to solve this issue by using
`vars` to replace the binary name and version with one that
is available for both arm64 and x86_64 platforms, satisfying
the dependencies when `gclient sync` runs.

It will not affect the build as `reclient` is not used.